Remove schema file
